generators/app/templates/python/requirements_api.txt
viam-sdk
<%= api_name_lower %>_api @ git+https://github.com/<%= api_namespace %>/<%= api_name_lower %>-api.git@main